added Fidelity types to schema module
This commit addresses (#998) and (#996) and corresponds to
1.) in @wchargin's suggested implementation plan.

The following adds a `Fidelity` type to Mirror.

`Fidelity` is either:
    - `FAITHFUL`
    - `UNFAITHFUL`

An `UNFAITHFUL` type has an `actualTypenames` property that lists
all the types of objects that can be returned when the field is
queried.

`fidelity` thus is added to Objects that are prone to inconsistent
typenames. These are `NodeFieldType` and `NestedFieldType`.

If any schema contains a Node or Nested Node  with a field that
contains an `UNFAITHFUL` `fidelity` type, a `Not Yet Implemented`
Error is thrown.

Test Plan:

A test has been added to the constructor to check that an error
is thrown given a schema that contains unfaithful fields.

Added a function for creating an unfaithful github schema,
`buildGithubSchemaUnfaithful`.

Added tests to `schema.test.js` to check validity of Unfaithful
Fidelity and duplicate `actualTypenames`