Improve PageRank performance
Based on profiling, about 30% of the time spent computing timeline
cred for SourceCred was spent in the `computeDelta` function. However,
the vast majority of the time when we were computing the delta, we
actually just needed to know whether any delta was above the maximum.
Rather than iterating over the whole distribution, we can short-circuit
as soon as we find any such delta.

This commit implements this change, and results in about a 30% speedup
for computing timelines: for sourcecred/sourcecred, the total time on my
machine (in browser) went from 14.9s to 10.4s. Profiling reveals that
the time spent in computeDelta is now negligible.

Test plan: Existing unit tests pass. I don't think a new one is needed.