Attempt to test for failure #1226
This commit attempts to modify sharness/test_build_static_site.t so as
to catch the error described in #1226. Unfortunately, the test does not
work as intended: the test passes even when the script is crashing.

The reason is that the exit trap in build_static_site.sh traps the
failure, and as such the script returns with exit code 0. Based on
googling, it should be possible to catch the actual exit code and
re-exit with it during the cleanup trap, and this indeed works for
"regular" exit codes, but not for the `1` that comes from referencing
an unset variable.

See the included script `problem.sh` for a reproduction.